Adjust ListMembersFromFamilyGroupResponse return type and other minor fixes
