+++ composeApp/src/commonMain/composeResources/values-en/strings.xml
@@ -3,6 +3,7 @@
     <string name="app_name">Family Vault</string>
     <string name="app_icon_alt">FamilyVault icon</string>
     <string name="loading">Loading...</string>
+    
 
     <!-- Icon descriptions -->
     <string name="info_icon_alt">Information icon</string>
