+++ composeApp/src/commonMain/composeResources/values-pl/strings.xml
@@ -3,6 +3,7 @@
     <string name="app_name">Family Vault</string>
     <string name="app_icon_alt">Ikona aplikacji Family Vault</string>
     <string name="loading">Ładowanie...</string>
+    
 
     <!-- Icon descriptions -->
     <string name="info_icon_alt">Ikona informacji</string>
