+++ composeApp/src/commonMain/kotlin/com/github/familyvault/backend/client/IFamilyVaultBackendClient.kt
@@ -1,6 +1,6 @@
 package com.github.familyvault.backend.client
 
-import com.github.familyvault.backend.requests.AddMemberToFamilyRequest
+import com.github.familyvault.backend.requests.AddMemberToFamilyGroupRequest
 import com.github.familyvault.backend.requests.CreateFamilyGroupRequest
 import com.github.familyvault.backend.requests.ListMembersFromFamilyGroupRequest
 import com.github.familyvault.backend.responses.CreateFamilyGroupResponse
@@ -10,8 +10,8 @@ import com.github.familyvault.backend.responses.PrivMxSolutionIdResponse
 interface IFamilyVaultBackendClient {
     suspend fun getSolutionId(): PrivMxSolutionIdResponse
     suspend fun createFamilyGroup(req: CreateFamilyGroupRequest) : CreateFamilyGroupResponse
-    suspend fun addGuardianToFamilyGroup(req: AddMemberToFamilyRequest)
-    suspend fun addMemberToFamilyGroup(req: AddMemberToFamilyRequest)
-    suspend fun addGuestToFamilyGroup(req: AddMemberToFamilyRequest)
+    suspend fun addGuardianToFamilyGroup(req: AddMemberToFamilyGroupRequest)
+    suspend fun addMemberToFamilyGroup(req: AddMemberToFamilyGroupRequest)
+    suspend fun addGuestToFamilyGroup(req: AddMemberToFamilyGroupRequest)
     suspend fun listMembersOfFamilyGroup(req: ListMembersFromFamilyGroupRequest): ListMembersFromFamilyGroupResponse
 }
