+++ composeApp/src/commonMain/kotlin/com/github/familyvault/backend/requests/AddMemberToFamilyGroupRequest.kt
@@ -1,0 +1,6 @@
+package com.github.familyvault.backend.requests
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddMemberToFamilyGroupRequest (val contextId: String, val userId: String, val userPubKey: String)
