--- composeApp/src/commonMain/kotlin/com/github/familyvault/backend/requests/AddMemberToFamilyRequest.kt
@@ -1,6 +1,0 @@
-package com.github.familyvault.backend.requests
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AddMemberToFamilyRequest (val contextId: String, val userId: String, val userPubKey: String)
