+++ composeApp/src/commonMain/kotlin/com/github/familyvault/backend/responses/ListMembersFromFamilyGroupResponse.kt
@@ -4,4 +4,4 @@ import com.github.familyvault.models.FamilyMember
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ListMembersFromFamilyGroupResponse (val members: MutableList<FamilyMember>)
+data class ListMembersFromFamilyGroupResponse (val members: List<FamilyMember>)
