+++ composeApp/src/commonMain/kotlin/com/github/familyvault/models/FamilyGroupMemberPermissionGroup.kt
@@ -1,0 +1,11 @@
+package com.github.familyvault.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class FamilyGroupMemberPermissionGroup(val value: Int) {
+    @SerialName("0") Guardian(0),
+    @SerialName("1") Member(1),
+    @SerialName("2") Guest(2);
+}
