+++ composeApp/src/commonMain/kotlin/com/github/familyvault/models/FamilyMember.kt
@@ -1,6 +1,5 @@
 package com.github.familyvault.models
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -12,12 +11,4 @@ data class FamilyMember(
 ){
     val fullname: String
         get() = "$firstname $surname"
-}
-
-
-@Serializable
-enum class FamilyGroupMemberPermissionGroup(val value: Int) {
-    @SerialName("0") Guardian(0),
-    @SerialName("1") Member(1),
-    @SerialName("2") Guest(2);
 }
