+++ composeApp/src/commonMain/kotlin/com/github/familyvault/services/FamilyGroupService.kt
@@ -3,7 +3,7 @@ package com.github.familyvault.services
 import com.github.familyvault.AppConfig
 import com.github.familyvault.backend.client.FamilyVaultBackendClient
 import com.github.familyvault.backend.client.IPrivMxClient
-import com.github.familyvault.backend.requests.AddMemberToFamilyRequest
+import com.github.familyvault.backend.requests.AddMemberToFamilyGroupRequest
 import com.github.familyvault.backend.requests.CreateFamilyGroupRequest
 import com.github.familyvault.backend.requests.ListMembersFromFamilyGroupRequest
 import com.github.familyvault.models.FamilyMember
@@ -33,7 +33,7 @@ class FamilyGroupService(
             CreateFamilyGroupRequest(familyGroupName, familyGroupDescription ?: "Test description")
         ).contextId
         familyVaultBackendProxy.addGuardianToFamilyGroup(
-            AddMemberToFamilyRequest(
+            AddMemberToFamilyGroupRequest(
                 contextId,
                 username,
                 pairOfKeys.publicKey
